+++ packages/facility-server/app/sync/initTimesync.js
@@ -1,0 +1,14 @@
+import config from '../config';
+import { create as createTimesync } from 'timesync';
+
+export const initTimesync = ({ models, centralServer }) => {
+  const timesync = createTimesync({
+    server: `${centralServer.host}/timesync`,
+    interval: config.timesync.interval,
+
+  });
+  timesync.on('change', (offset) => {
+    models.LocalSystemFact.set('timesyncOffset', offset);
+  });
+  return timesync;
+};
